refactor(ui): move store methods to a separate file
* improve reusability
* reduce code repetition
